Fixed improper stream frac usage:
The mixing fractions computed using the COSI weights were being used
improperly.  The COSI weight is an atom-based weight and so the fractions
calculated from it are correspondingly atom fractions.  Since Cyclus does all
material mixing on a mass-basis, the fractions need to first be converted to
mass-based fractions.  The AtomToMassFrac function was added to do this.  The
constraint converters for DRE, and all places in the code that use the
fractional interpolation based on weights (i.e. uses of HighFrac and LowFrac)
need to also convert the ratios to mass-based.  Added tests to check that this
is occurring correctly.